chore(op-program): move cross package out of op-supervisor
Relocates op-supervisor/supervisor/backend/cross to
op-program/client/interop/cross. The package contains cross-chain
hazard/cycle safety checks consumed by op-program's interop
consolidation (consolidate.go, interop_test.go) and by
op-supervisor's backend workers; only op-program will remain
once op-supervisor is deleted.

Pure relocation — import paths rewritten in the 3 external consumers
(op-program twice, op-supervisor backend once). No logic changes.
